chore: add base tests
